update to bid requirements

+++ app/helpers/helper.py
@@ -13,7 +13,7 @@ from app.crud import get_token
 from app.models.models import Token
 
 
-def convert_to_transaction_unit(payment_processor: AbstractPayment, value: int) -> int:
+def convert_to_transaction_unit(payment_processor: AbstractPayment, value: float) -> int:
 
     return payment_processor.unit_conversion(
         value=value,
